+++ AI2048Solver/heuristics.c
@@ -8,15 +8,15 @@
 
 #include "heuristics.h"
 
-int sum_heuristic(Board b) {
+double sum_heuristic(Board b) {
     int sum = 0;
     for (int r=0; r<BOARDSIZE; r++)
         for (int c=0; c<BOARDSIZE; c++)
             sum += bget(b, r, c);
-    return sum;
+    return (double)sum;
 }
 
-int squaresum_heuristic(Board b) {
+double squaresum_heuristic(Board b) {
     int sum = 0;
     for (int r=0; r<BOARDSIZE; r++) {
         for (int c=0; c<BOARDSIZE; c++) {
@@ -24,27 +24,27 @@ int squaresum_heuristic(Board b) {
             sum += this * this;
         }
     }
-    return sum;
+    return (double)sum;
 }
 
-int weights[BOARDSIZE][BOARDSIZE] = {
+double weights[BOARDSIZE][BOARDSIZE] = {
     { 0, 1, 2, 3},
     { 4, 5, 6, 7},
     { 8, 9,10,11},
     {12,13,14,15}
 };
-int weighted_sum1(Board b) {
-    int sum = 0;
+double weighted_sum1(Board b) {
+    double sum = 0.0;
     for (int r=0; r<BOARDSIZE; r++) {
         for (int c=0; c<BOARDSIZE; c++) {
             int this = bget(b, r, c);
-            sum += this * this * weights[r][c];
+            sum += (double)(this * this) * weights[r][c];
         }
     }
     return sum;
 }
 
-int weighted_sum2(Board b) {
+double weighted_sum2(Board b) {
     int sum = 0;
     for (int r=0; r<BOARDSIZE; r++) {
         for (int c=0; c<BOARDSIZE; c++) {
@@ -52,15 +52,15 @@ int weighted_sum2(Board b) {
             sum += this * this * (r + c);
         }
     }
-    return sum;
+    return (double)sum;
 }
 
-int empty_blocks(Board b) {
+double empty_blocks(Board b) {
     PointList pl = open_spaces(b);
     if (pl == NULL) {
         return 0;
     }
     int num_blocks = pl->N;
     pl_free(pl);
-    return num_blocks;
+    return (double)num_blocks;
 }
